+++ CHANGELOG.md
@@ -101,6 +101,9 @@
     `HTTP::Response` with which it was initialized -- no longer supports parsing
     data from external sources.
 - `Options` -- Rewritten with renamed option names and grouped relevant options together.
+- `AuditStore`
+    - `#save` -- Updated to store a compressed `Marshal` dump of the instance.
+    - `.load` -- Updated to load the new `#save` format.
 - Reports
     - Removed `metareport`.
 - Plugins